Split WithOffset into a separate file
